refactor(config): separate ConfigParser out from ConfigManager
