+++ src/config/config_types.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+namespace vkShade
+{
+    enum class ConfigError
+    {
+        KeyNotFound,
+        SectionNotFound,
+        FileNotFound,
+        ParseError
+    };
+} // namespace vkShade
